refactor(events): split system events into modules
