+++ apps/backend/app/domains/system/events/__init__.py
@@ -1,0 +1,27 @@
+from __future__ import annotations
+
+from .bus import EventBus, get_event_bus, register_handlers
+from .handlers import handlers
+from .models import (
+    EVENT_METRIC_NAMES,
+    AchievementUnlocked,
+    NodeArchived,
+    NodeCreated,
+    NodeEventBase,
+    NodePublished,
+    NodeUpdated,
+)
+
+__all__ = [
+    "NodeEventBase",
+    "NodeCreated",
+    "NodeUpdated",
+    "NodePublished",
+    "NodeArchived",
+    "AchievementUnlocked",
+    "EventBus",
+    "get_event_bus",
+    "register_handlers",
+    "handlers",
+    "EVENT_METRIC_NAMES",
+]
